Moving apis code to be under src, fix of yup schema (need to fix AutoForm as well)
