+++ pages/api/users/auth/logout.js
@@ -1,22 +1,4 @@
-import httpStatus from 'http-status'
-import jwt from 'jsonwebtoken'
-import responseJson from '@/utils/responseJson'
-import { TIME_UNITS } from '@/src/enumerators'
-import { strCookie } from '@/src/utils'
+import Logout from '@/src/api/users/auth/logout'
 
-const logout = (req, res) => {
-
-    const token = jwt.sign(
-        {},
-        process.env.JWT_SECRET, {
-            expiresIn: TIME_UNITS.YEAR * 120 / TIME_UNITS.SECOND,
-        }
-    )
-    const cookieString = strCookie('auth', token)
-
-    res.setHeader('location', '/')
-    res.setHeader('Set-Cookie', cookieString)
-    res.status(httpStatus.OK).json(responseJson(true, {}, httpStatus.OK))
-}
 // noinspection JSUnusedGlobalSymbols
-export default logout
+export default Logout
