+++ pages/api/users/index.js
@@ -1,14 +1,3 @@
-import httpStatus from 'http-status'
-import responseJson from '@/utils/responseJson'
-import { addUserToApi } from '@/src/utils'
+import Users from '@/src/api/users'
 
-const getUser = (res, user) => {
-    res.status(httpStatus.OK).json(responseJson(true, { user: user }, httpStatus.OK))
-}
-const users = (req, res, user) => {
-    switch (req.method) {
-        default:
-            getUser(res, user)
-    }
-}
-export default addUserToApi(users)
+export default Users
