+++ src/api/users/getUser.js
@@ -1,0 +1,8 @@
+import httpStatus from 'http-status'
+import responseJson from '@/utils/responseJson'
+
+const getUser = async (res, user) => {
+    res.status(httpStatus.OK).json(responseJson(true, { user: user }, httpStatus.OK))
+}
+
+export default getUser
