+++ src/components/AutoForm/AutoForm.module.scss
@@ -7,14 +7,13 @@
     grid-column-end: -3;
     grid-row-start: -1;
   }
-  &__error {
+  .error {
     color: var(--color-error);
     display: inline-block;
     font-size: var(--font-size-regular);
     font-weight: bold;
     text-align: var(--text-align-right);
   }
-
   grid-column-gap: var(--form-grid-column-gap);
   grid-row-gap: var(--form-grid-row-gap);
 }
